requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo
scikit-learn
dill
pyaml
-e.